reduced interface scope
