+++ beacon/validator/types.go
@@ -33,6 +33,7 @@ import (
 	"github.com/berachain/beacon-kit/primitives/eip4844"
 	"github.com/berachain/beacon-kit/primitives/math"
 	"github.com/berachain/beacon-kit/primitives/transition"
+	"github.com/berachain/beacon-kit/state-transition/core"
 	statedb "github.com/berachain/beacon-kit/state-transition/core/state"
 	depositdb "github.com/berachain/beacon-kit/storage/deposit"
 )
@@ -178,7 +179,7 @@ type StateProcessor interface {
 	) (transition.ValidatorUpdates, error)
 	// Transition performs the core state transition.
 	Transition(
-		ctx transition.ReadOnlyContext,
+		ctx core.ReadOnlyContext,
 		st *statedb.StateDB,
 		blk *ctypes.BeaconBlock,
 	) (transition.ValidatorUpdates, error)
