+++ node-core/components/interfaces.go
@@ -39,6 +39,7 @@ import (
 	"github.com/berachain/beacon-kit/primitives/eip4844"
 	"github.com/berachain/beacon-kit/primitives/math"
 	"github.com/berachain/beacon-kit/primitives/transition"
+	"github.com/berachain/beacon-kit/state-transition/core"
 	statedb "github.com/berachain/beacon-kit/state-transition/core/state"
 	"github.com/berachain/beacon-kit/storage/block"
 	depositdb "github.com/berachain/beacon-kit/storage/deposit"
@@ -331,7 +332,7 @@ type (
 		) (transition.ValidatorUpdates, error)
 		// Transition performs the core state transition.
 		Transition(
-			ctx transition.ReadOnlyContext,
+			ctx core.ReadOnlyContext,
 			st *statedb.StateDB,
 			blk *ctypes.BeaconBlock,
 		) (transition.ValidatorUpdates, error)
