+++ testing/state-transition/state-transition.go
@@ -104,7 +104,7 @@ func SetupTestState(t *testing.T, cs chain.Spec) (
 	*TestStateProcessorT,
 	*TestBeaconStateT,
 	*depositstore.KVStore,
-	transition.ReadOnlyContext,
+	core.ReadOnlyContext,
 ) {
 	t.Helper()
